split into main and notebook
