+++ index.js
@@ -4,4 +4,4 @@
 const path = require('path');
 
 const appRoot = path.join(__dirname);
-require('electron-compile').init(appRoot, './app');
+require('electron-compile').init(appRoot, './main');
