+++ notebook/reducers/app.js
