+++ package.json
@@ -9,7 +9,7 @@
     "test:renderer": "electron-mocha --renderer --compilers js:babel-core/register 'test/renderer/**/*.js'",
     "test:main": "electron-mocha --compilers js:babel-core/register 'test/main/**/*.js'",
     "test:watch": "watch 'npm run test' test/",
-    "start": "electron ."
+    "start": "electron index.js"
   },
   "repository": {
     "type": "git",
