+++ test/renderer/store/store_spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import createStore from '../../../src/store';
+import createStore from '../../../notebook/store';
 
 describe('createStore', () => {
   it('sets up our store model', (done) => {
